backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
httpx==0.26.0
sentence-transformers==2.5.1
python-dotenv==1.0.0
python-multipart==0.0.9
